requirements.txt
torch>=2.0.0
numpy
pandas
scikit-learn
matplotlib
tqdm
einops
pillow
albumentations
scipy
